switch logging to Zap and refactor to pass lint checks
Previously, we were using the Google Logging package which was logging to standard output even when it wasn't supposed to. I switched all logging to Uber's Zap package, which seems to be an industry standard for logging.
I also refactored all the files not up to standard passing Trunk's lint checks. I would have done this in a seperate commit, though the build will fail if trunk doesn't pass on all modified files. I decided that it would be better to resolve ALL of the trunk issues at once.